Adding PA 4
This programming assignment required me to construct a Fitness App program (C++). This will create a 7 day basic fitness application in C++, that allows the user of the application to manually edit “diet” and “exercise” plans. For this application you will need to create three classes: DietPlan, ExercisePlan, and FitnessAppWrapper.